Changed Demand Set h
Setting all variables to be a function of 'h'. This is due to an issue brought up today by Euan and Alex where the demand script may not match the data. By being able to run the Demand script for any value of 'h' (chosen by the user), then this issue should be minimised.